Quiz 3 - hasil nilai rata-rata dengan komanya

+++ main.go
@@ -6,6 +6,7 @@ func main() {
 
 	// hitung rata2
 	// scores := [8]int{100, 80, 75, 92, 70, 93, 88, 67}
+	// hasil harus beserta komanya
 	scores := [8]int{100, 80, 75, 92, 70, 93, 88, 67}
 
 	var total int
@@ -14,7 +15,7 @@ func main() {
 	}
 
 	length := len(scores)
-	average := total / length
+	average := float64(total) / float64(length)
 
 	fmt.Println(average)
